requirements.txt
anyio==3.6.1
APScheduler==3.9.1
autopep8==1.7.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
peewee==3.15.2
pycodestyle==2.9.1
python-dotenv==0.21.0
python-telegram-bot==20.0a4
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
toml==0.10.2
tornado==6.2
tradingview-ta==3.2.10
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
